requirements.txt
flask==3.0.0
groq
python-dotenv==1.0.0
fpdf2==2.8.2
requests==2.32.3